requirements.txt
asgiref==3.4.1
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.3
Django==4.0.1
django-clearcache==1.2.1
django-ratelimit==3.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.16.0
jmespath==0.10.0
numpy==1.26.0
pathspec==0.5.9
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
requests
sqlparse==0.4.2
termcolor==1.1.0
wcwidth==0.1.9
python-dotenv~=0.19.2
django-constance
django-constance[database]
django-allauth
aiocfscrape
cloudscraper~=1.2.64
django-cors-headers
feedparser~=6.0.10
django-ebhealthcheck
py-cord
python-dateutil~=2.8.0
humanize
tzlocal
pympler
urllib3==1.26.6
boto3==1.34.19
botocore==1.34.19

